+++ Dockerfile
@@ -9,11 +9,20 @@ ENV \
     COMMAND=${COMMAND} \
     DIRECTORY=${DIRECTORY}
 
-WORKDIR /go/bin/${DIRECTORY}
+WORKDIR /go/bin/lib
 
 RUN apk add --no-cache git
 
-COPY ${DIRECTORY}/lib /go/bin/${DIRECTORY}/lib
+COPY lib/go.mod /go/bin/lib
+COPY lib/go.sum /go/bin/lib
+RUN go mod download
+
+RUN go install github.com/jackstockley89/github-actions/lib/github@api
+
+COPY lib/${DIRECTORY} /go/bin/lib/${DIRECTORY}
+
+WORKDIR /go/bin/${DIRECTORY}
+
 COPY ${DIRECTORY}/go.mod /go/bin/${DIRECTORY}
 COPY ${DIRECTORY}/go.sum /go/bin/${DIRECTORY}
 RUN go mod download
