+++ github/version-check/version-check.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/jackstockley89/github-actions/github/lib"
+	lib "github.com/jackstockley89/github-actions/lib/github"
 	githubaction "github.com/sethvargo/go-githubactions"
 )
