+++ go.work
@@ -2,4 +2,6 @@ go 1.20
 
 use ./github
 
+use ./lib
+
 use ./aws
