+++ lib/aws/resource-clients.go
